Fixes #8
Was calling .trim() on strings when getting command input. Bad idea as they could be null anyway.
Wasn't checking the inputs early enough to determine what the attribute should be set to (ie update or keep the same). This has been solved in two steps.
step 1. Determine if the attribute is null, if so set to the current attribute, else set to new attribute. Store result in temp variable.
Step 2. Compare temp variable against current attribute, if not equal then update the attribute.